Add files via upload
The output of a to-do list application will vary depending on the actions taken by the user within the application. Here's a general overview of the program's behavior and potential outputs:

When the program starts, it presents a menu of options to the user, such as:

Add a task
View tasks
Delete a task
Exit
If the user chooses to add a task, they will be prompted to enter the task description. The task will then be added to the list.

If the user chooses to view tasks, the program will display all the tasks currently in the list.

If the user chooses to delete a task, the program will display the tasks along with their corresponding numbers. The user will then be prompted to enter the number of the task they want to delete.